Correction d'un bug dans Queue + changement au niveau du type de size

+++ stack.h
@@ -7,7 +7,7 @@
 
 typedef struct Stack{
 
-    unsigned int size;
+    unsigned long size;
     Tweet *top;
 
 }Stack;
